praticando Orientação a objeto

+++ screenmatch/src/br/com/ferpioli/screenmatch/modelos/Audio.java
@@ -22,7 +22,7 @@ public class Audio {
         return totalCurtidas;
     }
 
-    public double getClassificacao() {
+    public int getClassificacao() {
         return classificacao;
     }
